Create character conversion.py
This Python 3 program converts alphanumeric characters into patterns made out of non-alphanumeric characters, like "*" in this case. Currently working on the program to include as many alphanumeric characters as I can. Further, I plan on making a program to basically a sentence or even a paragraph, and it'll be converted into a pattern of characters.